git-fsck-cache.c: check commit objects more carefully
We historically used to be very careful in fsck-cache, but when it was
re-written to use "parse_object()" instead of parsing everything by
hand, it lost a bit of the checks.  This, together with the previous
commit, should make it do more proper commit object syntax checks.

Also add a "--strict" flag, which warns about the old-style "0664" file
mode bits, which shouldn't exist in modern trees, but that happened
early on in git trees and that the default git-fsck-cache thus silently
accepts.

Signed-off-by: Linus Torvalds <torvalds@osdl.org>

Former-commit-id: de2eb7f694da6396828e52588c9807a948a039a0