[PATCH] git-commit-script fix for degenerated merge
If merging results in an unchanged tree, git-commit-script should not
complain that there's nothing to commit.

Also, add "[--all]" to usage().

[jc: usually there is no reason to record an unchanging merge,
but this code path is triggered only when there is a nontrivial
merge that needed to be resolved by hand, and we should be able
to record the fact that these two tree heads are dealt with as a
regular two-parent commit in order to help later merges.]

Signed-off-by: Junio C Hamano <junkio@cox.net>

Former-commit-id: eca35ecdcaa1448de8cc9ec64590f47a9012264a