Better graph line details display and expand history coverage.
Now the history remembers when we have clicked on a graph line
and when we have asked for a diff between two commits, as well
as when we have displayed a commit.

The display when you click on a graph line now uses clickable
SHA1 IDs instead of the embedded "Go" buttons.  Also made the
IDs clickable in the header for a diff between two commits.


Former-commit-id: fa4da7b32e1197fa5032db6a2de7749868253f8b