Teach git push .git/branches shorthand
Although it is uncertain if we would keep .git/branches for
long, the shorthand stored there can be used for pushing if it
is host:path/to/git format, so let's make use of it.  This does
not use git-parse-remote because that script will be rewritten
quite a bit for updated pulling.

Signed-off-by: Junio C Hamano <junkio@cox.net>

Former-commit-id: 46b1c7c03c3ae3bf17aa87bf9ae2f9fb38d08789