Update rev-parse flags list.
I haven't audited the rev-parse users, but I am having a feeling
that many of them would choke when they expect a couple of SHA1
object names and malicious user feeds them "--max-count=6" or
somesuch to shoot himself in the foot.  Anyway, this adds a
couple of missing parameters that affect the list of revs to be
returned from rev-list, not the flags that affect how they are
presented by rev-list.  I think that is the intention, but I am
not quite sure.

Signed-off-by: Junio C Hamano <junkio@cox.net>

Former-commit-id: 5ccfb758b0e69b77b0ad830cc35d7770b08589eb