[PATCH] "Child" information in commit window - and cleanups
This adds "Child: " lines to the commit window, which tells what children
a commit has.

It also cleans things up: it marks the text widget as no-wrap, which means
that it doesn't need to truncate the commit description arbitrarily by
hand. Also, the description itself is now done by a common helper routine
that handles both the parent and the children.

Signed-off-by: Linus Torvalds <torvalds@osdl.org>
Signed-off-by: Paul Mackerras <paulus@samba.org>

Former-commit-id: b1ba39e7e80cdee3e7c1c80334cd9aa87b9b7b69