Add cheap local clone '-s' flag to git-clone-script
Using the $GIT_OBJECT_DIRECTORY/info/alternates mechanism,
create a new repository that borrows objects from the original
repository when --shared flag is given in addition to --local.

It is worth pointing out that the "cloned" repository depends on
the original repository, so this should be used only when you
can reasonably trust that the original repository would not
disappear without your knowing.

Signed-off-by: Junio C Hamano <junkio@cox.net>

Former-commit-id: aae4f42c45704986210ea2a30bd31afba1008275