[PATCH] clean up git script
Makes git work with a pure POSIX shell (tested with bash --posix and ash).
Right now git causes ash to choke on the redundant shift on line two.

Reduces the number of system calls git makes just to do a usage
statement from 22610 to 1122, and the runtime for same from 349ms to
29ms on my x86 Linux box.

Presents a standard usage statement, and pretty prints the available
commands in a form that does not scroll off small terminals.

[jc: while shifting when $# was zero was a bug, the original
patch failed to shift when it needs to, which I fixed up.]

Signed-off-by: Amos Waterland <apw@rossby.metr.ou.edu>
Signed-off-by: Junio C Hamano <junkio@cox.net>

Former-commit-id: b0c6f8d767387d39058648b4e3a04518350f789c