git-format-patch fixes.
Introduces --keep-subjects flag to tell it not to munge the
first line of the commit message.  Running "git applymbox" on
the output from "git format-patch -m -k" would preserve the
original commit information better this way.

At the same time, prefix Subject: on the first line of the
commit, to help people cut&copy.

Signed-off-by: Junio C Hamano <junkio@cox.net>

Former-commit-id: af5260ee785b8418c7319a84b1d1a0d26e770ed0