[PATCH] Fix git-format-patch-script to handle empty messages
In case of a commit with an empty message there is no
mandatory empty line between headers and body

[jc: This makes --mbox output valid even when the commit message does
not have anything but its first line, which the one I wrote botched.
One side-effect is that it adds an extra blank line at the end even if
it has more than one lines, which will be eaten by the receiving end.
As Marco says, this is a stop-gap measure.  This script needs to be
split into two, one that gets the format specifier and a commit ID to
write to its standard output, and another that drives that one reading
from rev-list.  I'll fix things properly when that happens by
rewriting the former part in Perl or something more reasonable than
the current shell, sed and grep mishmash.]

Signed-off-by: Marco Costalba <mcostalba@yahoo.it>
Signed-off-by: Junio C Hamano <junkio@cox.net>

Former-commit-id: cc5625a5e2e10a94560347333c43b7032c2566f3