Added re-read refs command, and display all refs.
These are features requested by Junio.  Any plain file under .git/refs
whose contents start with 40 hex characters is taken as a reference
and displayed like a head but with a light blue background (unless it
is in .git/refs/tags or .git/refs/heads, in which case it is displayed
as before).  There is now a "Reread references" menu item in the File
menu which re-reads all the plain files under .git/refs and redisplays
any references that have changed.


Former-commit-id: f1d83ba34ccb4c7fa8c488a669eeaaba1adc68e0