mailinfo and applymbox updates
This attempts to minimally cope with a subset of MIME "features" often
seen in patches sent to our mailing lists.  Namely:

 - People's name spelled in characters outside ASCII (both on From:
   header and the signed-off-by line).

 - Content-transfer-encoding using quoted-printable (both in
   multipart and non-multipart messages).

These MIME features are detected and decoded by "git mailinfo".

Optionally, with the '-u' flag, the output to .info and .msg is
transliterated from its original chaset to utf-8.  This is to
encourage people to use utf8 in their commit messages for
interoperability.

Applymbox accepts additional flag '-u' which is passed to mailinfo.

Signed-off-by: Junio C Hamano / 濱野 純 <junkio@cox.net>

Former-commit-id: d4a9ce78f6d516568bfca4d808033d9e8254c33f