git-repack-script: Add option to repack all objects.
This originally came from Frank Sorenson, but with a bit of rework to
allow future enhancements without changing the external interface for
pack pruning part.

With the '-a' option, all objects in the current repository are packed
into a single pack.  When the '-d' option is given at the same time,
existing packs that were made redundant by this round of repacking are
deleted.

Since we currently have only two repacking strategies, one with '-a'
(everything into one) and the other without '-a' (incrementally pack
only the unpacked ones), the '-d' option is meaningful only when used
with '-a'; it removes the packs existed before we did the "everything
into one" repacking.  At least for now.

Signed-off-by: Junio C Hamano <junkio@cox.net>
Acked-by: Frank Sorenson <frank@tuxrocks.com>
(cherry picked from bfed505327e31221d8de796b3af880bad696b149 commit)

Former-commit-id: d5acdcf18e255308f5f16ae9e7fe48bf21ff68c0