Fast-path 'update-index --refresh' a bit.
If the length in the stat information does not match what is recorded
in the index, there is no point rehashing the contents to see if the
index entry can be refreshed.

We need to be a bit careful.  Immediately after read-tree or
checkout-index without -u, ce_size is set to zero and does not match
the length of the blob that is recorded, and we need to actually look
at the contents to see if it has been changed.

Signed-off-by: Junio C Hamano <junkio@cox.net>

Former-commit-id: 1991b223c05d45d2a915842990b2a7c4819dcbcf