Clean mail files after dealing with them.
When you are applying 200 mails in sequence, .dotest/ directory
will be littered with many messsages, and when the patch in one
of them fails to apply, it is not obvious which message was
being processed.  Remove the one that has been already dealt
with, so that the last failed one is found typically as the
lowest numbered split message.

Signed-off-by: Junio C Hamano <junkio@cox.net>

Former-commit-id: 4ebe63dfe6f2a8c652ecd8dfe6ab8a77b6c82b2f